pyproject.toml
[project]
name = "ry-tool"
version = "0.3.1"
description = "Pure YAML command orchestrator - CI/CD for humans"
readme = "README.md"
authors = [
    { name = "Fredrik Angelsen", email = "fredrikangelsen@gmail.com" }
]
requires-python = ">=3.12"
dependencies = [
    "pyyaml>=6.0",
]

[project.scripts]
ry = "ry_tool.__main__:main"

[build-system]
requires = ["uv_build>=0.8.13,<0.9.0"]
build-backend = "uv_build"
